+++ src/logger.js
@@ -1,20 +1,29 @@
-import { inAgencyRun } from './agency'
+import { isInAgency } from './agency'
+
+const curLevel = typeof process.env.ANTARES_LOG_LEVEL === 'undefined'
+  ? 2
+  : Number(process.env.ANTARES_LOG_LEVEL)
 
 export const logger = {
-  log: (msg, opts = {}) => {
-    const xmsg = opts.xform ? opts.xform(msg) : msg
-    const toPrint = (typeof xmsg === 'string') ? xmsg : JSON.stringify(xmsg)
-    const inAgency = opts.inAgency ? opts.inAgency : 'any'
+  log: (
+    msg,
+    { xform = x => x, inAgency = 'any', level = 2, newSection, prefix }
+  ) => {
+    const xmsg = xform(msg)
+    const toPrint = typeof xmsg === 'string' ? xmsg : JSON.stringify(xmsg)
 
-    inAgencyRun(inAgency, () => {
-      const prefix = opts.prefix ? (opts.prefix + '> ') : ''
-      if (opts.newSection) {
-        console.log('  --------------  ')
-      }
-      console.log(prefix + toPrint)
-    })
+    if (!isInAgency(inAgency)) return
+    if (level > curLevel) return
+
+    const prefixText = prefix ? prefix + '> ' : ''
+    if (newSection) {
+      console.log('  --------------  ')
+    }
+    console.log(prefixText + toPrint)
   },
-  debug: (...args) => logger.log(...args)
+  debug: (msg, opts) => logger.log(msg, { ...opts, level: 3 }),
+  warn: (msg, opts) => logger.log(msg, { ...opts, level: 1 }),
+  error: (msg, opts) => logger.log(msg, { ...opts, level: 0 })
 }
 
 export const ppAction = action => {